Prepared HookHandler for multithreading.
The only member which could be modified form another thread is the key
combination array. So I added a mutex which is locked when either thread
needs to access the array.

In the hook thread the the mutex waits only for a very short time 10ms. It
should be enough except in the case the application is bugged. Thanks to
that the other thread cannot block the hooke execution.

I think this duration could be extended without problems to 100ms as
nobody types more than 10 characters by second (I would like to know such
a person :-) ). But 10ms should be good on most PCs.