Splitted the execution into 2 threads.
The main thread is in charge of the initialisation and runs the hook.
The other thread is in charge of executing the required actions.

The communication between the two threads is done with a semaphore which
is incremented whenever a new shortcut is triggered and decremented when
the worker thread processes the event, and a queue to post shortcuts to be
executed. The queue is also synchronized with le lock key word.