Added a function to wait for modifiers release.
This is needed because when a process is starting or when the screen is
locked, the handler does not recieve key up events for the modifiers. This
can be very troublesome.
Waiting for the modifiers to be released before starting process or
locking the screen should not be a bad idea.